PICARD-2576: Moved toolbar customization to separate options page
This removes complexity from the general "User Interface" page and allows the options dialog to be resized to a smaller height, while still allow comfortable toolbar customization. This makes the options usable again on screens with only 768px height available.
