+++ picard/ui/options/dialog.py
@@ -65,6 +65,7 @@ from picard.ui.options import (  # noqa: F401 # pylint: disable=unused-import
     genres,
     interface,
     interface_colors,
+    interface_toolbar,
     interface_top_tags,
     maintenance,
     matching,
