+++ picard/ui/options/interface.py
@@ -39,19 +39,13 @@ from PyQt5.QtCore import QStandardPaths
 
 from picard.config import (
     BoolOption,
-    ListOption,
     TextOption,
     get_config,
 )
 from picard.const import UI_LANGUAGES
 from picard.const.sys import IS_MACOS
-from picard.util import (
-    icontheme,
-    strxfrm,
-)
+from picard.util import strxfrm
 
-from picard.ui import PicardDialog
-from picard.ui.moveable_list_view import MoveableListView
 from picard.ui.options import (
     OptionsPage,
     register_options_page,
@@ -75,66 +69,7 @@ class InterfaceOptionsPage(OptionsPage):
     SORT_ORDER = 80
     ACTIVE = True
     HELP_URL = '/config/options_interface.html'
-    SEPARATOR = '—' * 5
-    TOOLBAR_BUTTONS = {
-        'add_directory_action': {
-            'label': N_('Add Folder'),
-            'icon': 'folder'
-        },
-        'add_files_action': {
-            'label': N_('Add Files'),
-            'icon': 'document-open'
-        },
-        'cluster_action': {
-            'label': N_('Cluster'),
-            'icon': 'picard-cluster'
-        },
-        'autotag_action': {
-            'label': N_('Lookup'),
-            'icon': 'picard-auto-tag'
-        },
-        'analyze_action': {
-            'label': N_('Scan'),
-            'icon': 'picard-analyze'
-        },
-        'browser_lookup_action': {
-            'label': N_('Lookup in Browser'),
-            'icon': 'lookup-musicbrainz'
-        },
-        'save_action': {
-            'label': N_('Save'),
-            'icon': 'document-save'
-        },
-        'view_info_action': {
-            'label': N_('Info'),
-            'icon': 'picard-edit-tags'
-        },
-        'remove_action': {
-            'label': N_('Remove'),
-            'icon': 'list-remove'
-        },
-        'submit_acoustid_action': {
-            'label': N_('Submit AcoustIDs'),
-            'icon': 'acoustid-fingerprinter'
-        },
-        'generate_fingerprints_action': {
-            'label': N_("Generate Fingerprints"),
-            'icon': 'fingerprint'
-        },
-        'play_file_action': {
-            'label': N_('Open in Player'),
-            'icon': 'play-music'
-        },
-        'cd_lookup_action': {
-            'label': N_('Lookup CD...'),
-            'icon': 'media-optical'
-        },
-        'tags_from_filenames_action': {
-            'label': N_('Parse File Names...'),
-            'icon': 'picard-tags-from-filename'
-        },
-    }
-    ACTION_NAMES = set(TOOLBAR_BUTTONS.keys())
+
     options = [
         BoolOption("setting", "toolbar_show_labels", True),
         BoolOption("setting", "toolbar_multiselect", False),
@@ -148,24 +83,6 @@ class InterfaceOptionsPage(OptionsPage):
         BoolOption("setting", "starting_directory", False),
         TextOption("setting", "starting_directory_path", _default_starting_dir),
         TextOption("setting", "load_image_behavior", "append"),
-        ListOption("setting", "toolbar_layout", [
-            'add_directory_action',
-            'add_files_action',
-            'separator',
-            'cluster_action',
-            'separator',
-            'autotag_action',
-            'analyze_action',
-            'browser_lookup_action',
-            'separator',
-            'save_action',
-            'view_info_action',
-            'remove_action',
-            'separator',
-            'cd_lookup_action',
-            'separator',
-            'submit_acoustid_action',
-        ]),
     ]
 
     # Those are labels for theme display
@@ -220,12 +137,6 @@ class InterfaceOptionsPage(OptionsPage):
             self.ui.starting_directory_browse.setEnabled
         )
         self.ui.starting_directory_browse.clicked.connect(self.starting_directory_browse)
-        self.ui.add_button.clicked.connect(self.add_to_toolbar)
-        self.ui.insert_separator_button.clicked.connect(self.insert_separator)
-        self.ui.remove_button.clicked.connect(self.remove_action)
-        self.move_view = MoveableListView(self.ui.toolbar_layout_list, self.ui.up_button,
-                                          self.ui.down_button, self.update_action_buttons)
-        self.update_buttons = self.move_view.update_buttons
 
         if not OS_SUPPORTS_THEMES:
             self.ui.ui_theme_container.hide()
@@ -243,11 +154,8 @@ class InterfaceOptionsPage(OptionsPage):
         self.ui.filebrowser_horizontal_autoscroll.setChecked(config.setting["filebrowser_horizontal_autoscroll"])
         self.ui.starting_directory.setChecked(config.setting["starting_directory"])
         self.ui.starting_directory_path.setText(config.setting["starting_directory_path"])
-        self.populate_action_list()
-        self.ui.toolbar_layout_list.setCurrentRow(0)
         current_theme = UiTheme(config.setting["ui_theme"])
         self.ui.ui_theme.setCurrentIndex(self.ui.ui_theme.findData(current_theme))
-        self.update_buttons()
 
     def save(self):
         config = get_config()
@@ -286,11 +194,6 @@ class InterfaceOptionsPage(OptionsPage):
         config.setting["filebrowser_horizontal_autoscroll"] = self.ui.filebrowser_horizontal_autoscroll.isChecked()
         config.setting["starting_directory"] = self.ui.starting_directory.isChecked()
         config.setting["starting_directory_path"] = os.path.normpath(self.ui.starting_directory_path.text())
-        self.update_layout_config()
-
-    def restore_defaults(self):
-        super().restore_defaults()
-        self.update_buttons()
 
     def starting_directory_browse(self):
         item = self.ui.starting_directory_path
@@ -298,117 +201,6 @@ class InterfaceOptionsPage(OptionsPage):
         if path:
             path = os.path.normpath(path)
             item.setText(path)
-
-    def _get_icon_from_name(self, name):
-        return self.TOOLBAR_BUTTONS[name]['icon']
-
-    def _insert_item(self, action, index=None):
-        list_item = ToolbarListItem(action)
-        list_item.setToolTip(_('Drag and Drop to re-order'))
-        if action in self.TOOLBAR_BUTTONS:
-            # TODO: Remove temporary workaround once https://github.com/python-babel/babel/issues/415 has been resolved.
-            babel_415_workaround = self.TOOLBAR_BUTTONS[action]['label']
-            list_item.setText(_(babel_415_workaround))
-            list_item.setIcon(icontheme.lookup(self._get_icon_from_name(action), icontheme.ICON_SIZE_MENU))
-        else:
-            list_item.setText(self.SEPARATOR)
-        if index is not None:
-            self.ui.toolbar_layout_list.insertItem(index, list_item)
-        else:
-            self.ui.toolbar_layout_list.addItem(list_item)
-        return list_item
-
-    def _all_list_items(self):
-        return [self.ui.toolbar_layout_list.item(i).action_name
-                for i in range(self.ui.toolbar_layout_list.count())]
-
-    def _added_actions(self):
-        actions = self._all_list_items()
-        return set(action for action in actions if action != 'separator')
-
-    def populate_action_list(self):
-        self.ui.toolbar_layout_list.clear()
-        config = get_config()
-        for name in config.setting['toolbar_layout']:
-            if name in self.ACTION_NAMES or name == 'separator':
-                self._insert_item(name)
-
-    def update_action_buttons(self):
-        self.ui.add_button.setEnabled(self._added_actions() != self.ACTION_NAMES)
-
-    def add_to_toolbar(self):
-        display_list = set.difference(self.ACTION_NAMES, self._added_actions())
-        selected_action, ok = AddActionDialog.get_selected_action(display_list, self)
-        if ok:
-            list_item = self._insert_item(selected_action, self.ui.toolbar_layout_list.currentRow() + 1)
-            self.ui.toolbar_layout_list.setCurrentItem(list_item)
-        self.update_buttons()
-
-    def insert_separator(self):
-        insert_index = self.ui.toolbar_layout_list.currentRow() + 1
-        self._insert_item('separator', insert_index)
-
-    def remove_action(self):
-        item = self.ui.toolbar_layout_list.takeItem(self.ui.toolbar_layout_list.currentRow())
-        del item
-        self.update_buttons()
-
-    def update_layout_config(self):
-        config = get_config()
-        config.setting['toolbar_layout'] = self._all_list_items()
-        self._update_toolbar()
-
-    def _update_toolbar(self):
-        widget = self.parent()
-        while not isinstance(widget, QtWidgets.QMainWindow):
-            widget = widget.parent()
-        # Call the main window's create toolbar method
-        widget.create_action_toolbar()
-        widget.update_toolbar_style()
-        widget.set_tab_order()
-
-
-class ToolbarListItem(QtWidgets.QListWidgetItem):
-    def __init__(self, action_name, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.action_name = action_name
-
-
-class AddActionDialog(PicardDialog):
-    def __init__(self, action_list, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.setWindowModality(QtCore.Qt.WindowModality.WindowModal)
-
-        layout = QtWidgets.QVBoxLayout(self)
-        layout.setSizeConstraint(QtWidgets.QLayout.SizeConstraint.SetFixedSize)
-
-        # TODO: Remove temporary workaround once https://github.com/python-babel/babel/issues/415 has been resolved.
-        babel_415_workaround_list = []
-        for action in action_list:
-            babel_415_workaround = self.parent().TOOLBAR_BUTTONS[action]['label']
-            babel_415_workaround_list.append([_(babel_415_workaround), action])
-        self.action_list = sorted(babel_415_workaround_list)
-
-        self.combo_box = QtWidgets.QComboBox(self)
-        self.combo_box.addItems([label for label, action in self.action_list])
-        layout.addWidget(self.combo_box)
-
-        buttons = QtWidgets.QDialogButtonBox(
-            QtWidgets.QDialogButtonBox.StandardButton.Ok | QtWidgets.QDialogButtonBox.StandardButton.Cancel,
-            QtCore.Qt.Orientation.Horizontal, self)
-        buttons.accepted.connect(self.accept)
-        buttons.rejected.connect(self.reject)
-        layout.addWidget(buttons)
-
-    def selected_action(self):
-        return self.action_list[self.combo_box.currentIndex()][1]
-
-    @staticmethod
-    def get_selected_action(action_list, parent=None):
-        dialog = AddActionDialog(action_list, parent)
-        result = dialog.exec_()
-        selected_action = dialog.selected_action()
-        return (selected_action, result == QtWidgets.QDialog.DialogCode.Accepted)
 
 
 register_options_page(InterfaceOptionsPage)
