+++ picard/ui/ui_options_interface.py
@@ -10,7 +10,7 @@ from PyQt5 import QtCore, QtGui, QtWidgets
 class Ui_InterfaceOptionsPage(object):
     def setupUi(self, InterfaceOptionsPage):
         InterfaceOptionsPage.setObjectName("InterfaceOptionsPage")
-        InterfaceOptionsPage.resize(466, 735)
+        InterfaceOptionsPage.resize(466, 543)
         self.vboxlayout = QtWidgets.QVBoxLayout(InterfaceOptionsPage)
         self.vboxlayout.setObjectName("vboxlayout")
         self.groupBox = QtWidgets.QGroupBox(InterfaceOptionsPage)
@@ -88,51 +88,8 @@ class Ui_InterfaceOptionsPage(object):
         self.verticalLayout_2.setObjectName("verticalLayout_2")
         self.vboxlayout1.addWidget(self.ui_theme_container)
         self.vboxlayout.addWidget(self.miscellaneous_box)
-        self.customize_toolbar_box = QtWidgets.QGroupBox(InterfaceOptionsPage)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.MinimumExpanding, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.customize_toolbar_box.sizePolicy().hasHeightForWidth())
-        self.customize_toolbar_box.setSizePolicy(sizePolicy)
-        self.customize_toolbar_box.setObjectName("customize_toolbar_box")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.customize_toolbar_box)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.toolbar_layout_list = QtWidgets.QListWidget(self.customize_toolbar_box)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Expanding)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.toolbar_layout_list.sizePolicy().hasHeightForWidth())
-        self.toolbar_layout_list.setSizePolicy(sizePolicy)
-        self.toolbar_layout_list.setObjectName("toolbar_layout_list")
-        self.verticalLayout.addWidget(self.toolbar_layout_list)
-        self.edit_button_box = QtWidgets.QWidget(self.customize_toolbar_box)
-        self.edit_button_box.setObjectName("edit_button_box")
-        self.edit_box_layout = QtWidgets.QHBoxLayout(self.edit_button_box)
-        self.edit_box_layout.setContentsMargins(0, 0, 0, 0)
-        self.edit_box_layout.setObjectName("edit_box_layout")
-        self.add_button = QtWidgets.QToolButton(self.edit_button_box)
-        self.add_button.setObjectName("add_button")
-        self.edit_box_layout.addWidget(self.add_button)
-        self.insert_separator_button = QtWidgets.QToolButton(self.edit_button_box)
-        self.insert_separator_button.setObjectName("insert_separator_button")
-        self.edit_box_layout.addWidget(self.insert_separator_button)
-        spacerItem2 = QtWidgets.QSpacerItem(50, 20, QtWidgets.QSizePolicy.MinimumExpanding, QtWidgets.QSizePolicy.Minimum)
-        self.edit_box_layout.addItem(spacerItem2)
-        self.up_button = QtWidgets.QToolButton(self.edit_button_box)
-        icon = QtGui.QIcon.fromTheme(":/images/16x16/go-up.png")
-        self.up_button.setIcon(icon)
-        self.up_button.setObjectName("up_button")
-        self.edit_box_layout.addWidget(self.up_button)
-        self.down_button = QtWidgets.QToolButton(self.edit_button_box)
-        icon = QtGui.QIcon.fromTheme(":/images/16x16/go-down.png")
-        self.down_button.setIcon(icon)
-        self.down_button.setObjectName("down_button")
-        self.edit_box_layout.addWidget(self.down_button)
-        self.remove_button = QtWidgets.QToolButton(self.edit_button_box)
-        self.remove_button.setObjectName("remove_button")
-        self.edit_box_layout.addWidget(self.remove_button)
-        self.verticalLayout.addWidget(self.edit_button_box)
-        self.vboxlayout.addWidget(self.customize_toolbar_box)
+        spacerItem2 = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
+        self.vboxlayout.addItem(spacerItem2)
 
         self.retranslateUi(InterfaceOptionsPage)
         QtCore.QMetaObject.connectSlotsByName(InterfaceOptionsPage)
@@ -147,12 +104,6 @@ class Ui_InterfaceOptionsPage(object):
         InterfaceOptionsPage.setTabOrder(self.filebrowser_horizontal_autoscroll, self.starting_directory)
         InterfaceOptionsPage.setTabOrder(self.starting_directory, self.starting_directory_path)
         InterfaceOptionsPage.setTabOrder(self.starting_directory_path, self.starting_directory_browse)
-        InterfaceOptionsPage.setTabOrder(self.starting_directory_browse, self.toolbar_layout_list)
-        InterfaceOptionsPage.setTabOrder(self.toolbar_layout_list, self.add_button)
-        InterfaceOptionsPage.setTabOrder(self.add_button, self.insert_separator_button)
-        InterfaceOptionsPage.setTabOrder(self.insert_separator_button, self.up_button)
-        InterfaceOptionsPage.setTabOrder(self.up_button, self.down_button)
-        InterfaceOptionsPage.setTabOrder(self.down_button, self.remove_button)
 
     def retranslateUi(self, InterfaceOptionsPage):
         _translate = QtCore.QCoreApplication.translate
@@ -169,12 +120,3 @@ class Ui_InterfaceOptionsPage(object):
         self.filebrowser_horizontal_autoscroll.setText(_("Adjust horizontal position in file browser automatically"))
         self.starting_directory.setText(_("Begin browsing in the following directory:"))
         self.starting_directory_browse.setText(_("Browse..."))
-        self.customize_toolbar_box.setTitle(_("Customize Action Toolbar"))
-        self.add_button.setToolTip(_("Add a new button to Toolbar"))
-        self.add_button.setText(_("Add Action"))
-        self.insert_separator_button.setToolTip(_("Insert a separator"))
-        self.insert_separator_button.setText(_("Add Separator"))
-        self.up_button.setToolTip(_("Move selected item up"))
-        self.down_button.setToolTip(_("Move selected item down"))
-        self.remove_button.setToolTip(_("Remove button from toolbar"))
-        self.remove_button.setText(_("Remove"))
